% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.onLoad}
\alias{.onLoad}
\title{Load ufal.morphodita python package}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library name}

\item{pkgname}{Package name}
}
\description{
Load ufal.morphodita python package
}
