% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{download_models}
\alias{download_models}
\title{Download language models from LINDAT repository}
\usage{
download_models(lang = "CZ", dest_folder = "")
}
\arguments{
\item{lang}{Language of the model ("CZ" for Czech, "SK" for Slovak and "EN" for English)}

\item{dest_folder}{Destination folder where the language models should be downloaded to}
}
\value{
Vector with paths to available models (taggers and morphological dictionaries)
}
\description{
Download language models from LINDAT repository
}
