% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extract_conll_tags}
\alias{extract_conll_tags}
\title{Split CONLL tags into separate columns}
\usage{
extract_conll_tags(df)
}
\arguments{
\item{df}{Data.frame containing column `tag` with morphological tags}
}
\description{
Split CONLL tags into separate columns
}
\seealso{
https://ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html
}
