% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{extract_hm_tags}
\alias{extract_hm_tags}
\title{Split HM tags into separate columns}
\usage{
extract_hm_tags(df)
}
\arguments{
\item{df}{Data.frame containing column `tag` with morphological tags}
}
\description{
Split HM tags into separate columns
}
\seealso{
https://ufal.mff.cuni.cz/pdt/Morphology_and_Tagging/Doc/hmptagqr.html
}
