% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{load_morpho}
\alias{load_morpho}
\title{Load Morphological dictionary}
\usage{
load_morpho(src)
}
\arguments{
\item{src}{Path to the morphological dictionary}
}
\description{
Load Morphological dictionary
}
\examples{
\dontrun{
load_morpho("src/czech-morfflex-161115.dict")
}
}
