% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{morpho_analyze}
\alias{morpho_analyze}
\title{Analyze morphology of a word}
\usage{
morpho_analyze(tagger, x)
}
\arguments{
\item{tagger}{Morphodita tagger}

\item{x}{a word to analyze}
}
\description{
Analyze morphology of a word
}
