% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{morpho_generate}
\alias{morpho_generate}
\title{Generate forms of a given lemma}
\usage{
morpho_generate(tagger, lemma, tag_wildcard = NULL)
}
\arguments{
\item{tagger}{morphological tagger}

\item{lemma}{a lemma from which forms should be generated}

\item{tag_wildcard}{specification of a form that should be generated}
}
\description{
Generate forms of a given lemma
}
