% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{data}
\name{morphodita}
\alias{morphodita}
\title{Morphodita Python binding}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.
}
\usage{
morphodita
}
\description{
Morphodita Python binding
}
\keyword{datasets}
