% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{recode_tag}
\alias{recode_tag}
\title{Recode tag to factor}
\usage{
recode_tag(x, tags_df, tag_type)
}
\arguments{
\item{x}{tag to be recoded}

\item{tags_df}{List of Data.frames with values and descriptions of morphological tags}

\item{tag_type}{Abbreviation of a morphological tag}
}
\description{
Recode tag to factor
}
