% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{recode_tags}
\alias{recode_tags}
\title{Recode morphological tags}
\usage{
recode_tags(df, tags_df)
}
\arguments{
\item{df}{Data.frame with extracted tags}

\item{tags_df}{List of data.frames with values and descriptions of morphological tags}
}
\description{
Recode morphological tags
}
