% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAGS}
\alias{TAGS}
\title{Positional Tags}
\format{
A list with 13 data.frames each containing value and description of
        positional tags
\describe{
  \item{POS}{Part of Speech}
  \item{SUBPOS}{Detailed Part of Speech}
  \item{GENDER}{Gender}
  ...
}
}
\source{
\url{https://ufal.mff.cuni.cz/pdt/Morphology_and_Tagging/Doc/hmptagqr.html}
}
\usage{
TAGS
}
\description{
Reference of Czech HM Morphology
}
\keyword{datasets}
