% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TAGS_EN}
\alias{TAGS_EN}
\title{Positional Tags - Penn Treebank Project}
\format{
data.frame of 36 rows and 2 variables
\describe{
  \item{Value}{Part of Speech tag}
  \item{Description}{Description of the tag}
}
}
\source{
\url{http://www.ling.upenn.edu/courses/Fall_2003/ling001/penn_treebank_pos.html}
}
\usage{
TAGS_EN
}
\description{
Reference of Penn Treebank Project Part of Speech tags
}
\keyword{datasets}
